+++ ranger-http/src/main/java/io/appform/ranger/http/servicefinder/HttpNodeDataSource.java
@@ -16,6 +16,7 @@
 package io.appform.ranger.http.servicefinder;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.base.Function;
 import com.google.common.base.Preconditions;
 import io.appform.ranger.core.model.NodeDataSource;
 import io.appform.ranger.core.model.Service;
@@ -24,15 +25,18 @@ import io.appform.ranger.core.util.FinderUtils;
 import io.appform.ranger.http.common.HttpNodeDataStoreConnector;
 import io.appform.ranger.http.config.HttpClientConfig;
 import io.appform.ranger.http.serde.HTTPResponseDataDeserializer;
-import java.util.Optional;
+import io.appform.ranger.http.utils.HttpClientUtils;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
-import okhttp3.HttpUrl;
-import okhttp3.Request;
+import org.apache.hc.client5.http.fluent.Executor;
+import org.apache.hc.client5.http.fluent.Request;
+import org.apache.hc.core5.net.URIBuilder;
+import org.apache.hc.core5.util.Timeout;
 
-import java.io.IOException;
-import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
 
 /**
  *
@@ -45,62 +49,46 @@ public class HttpNodeDataSource<T, D extends HTTPResponseDataDeserializer<T>> ex
     public HttpNodeDataSource(
             Service service,
             final HttpClientConfig config,
-            ObjectMapper mapper) {
-        super(config, mapper);
+            ObjectMapper mapper,
+            Executor executor) {
+        super(config, mapper, executor);
         this.service = service;
     }
 
     @Override
+    @SneakyThrows
     public Optional<List<ServiceNode<T>>> refresh(D deserializer) {
         Preconditions.checkNotNull(config, "client config has not been set for node data");
         Preconditions.checkNotNull(mapper, "mapper has not been set for node data");
         val url = String.format("/ranger/nodes/v1/%s/%s", service.getNamespace(), service.getServiceName());
 
         log.debug("Refreshing the node list from url {}", url);
-        val httpUrl = new HttpUrl.Builder()
-                .scheme(config.isSecure()
+
+        val httpUrl = new URIBuilder()
+                .setScheme(config.isSecure()
                         ? "https"
                         : "http")
-                .host(config.getHost())
-                .port(config.getPort() == 0
+                .setHost(config.getHost())
+                .setPort(config.getPort() == 0
                         ? defaultPort()
                         : config.getPort())
-                .encodedPath(url)
-                .build();
-        val request = new Request.Builder()
-                .url(httpUrl)
-                .get()
+                .setPath(url)
                 .build();
 
-        try (val response = httpClient.newCall(request).execute()) {
-            if (response.isSuccessful()) {
-                try (val body = response.body()) {
-                    if (null == body) {
-                        log.warn("HTTP call to {} returned empty body", httpUrl);
-                    } else {
-                        val bytes = body.bytes();
-                        val serviceNodesResponse = deserializer.deserialize(bytes);
-                        if(serviceNodesResponse.valid()){
-                            return Optional.of(FinderUtils.filterValidNodes(
-                                               service,
-                                               serviceNodesResponse.getData(),
-                                               healthcheckZombieCheckThresholdTime(service)));
-                        } else{
-                            log.warn("Http call to {} returned a failure response with response {}", httpUrl, serviceNodesResponse);
-                        }
-                    }
-                }
+        return HttpClientUtils.executeRequest(httpExecutor, Request.get(httpUrl), (Function<byte[], Optional<List<ServiceNode<T>>>>) responseBytes -> {
+            val serviceNodesResponse = deserializer.deserialize(responseBytes);
+            if (serviceNodesResponse.valid()) {
+                return Optional.of(FinderUtils.filterValidNodes(
+                        service,
+                        serviceNodesResponse.getData(),
+                        healthcheckZombieCheckThresholdTime(service)));
             } else {
-                log.warn("HTTP call to {} returned: {}", httpUrl, response.code());
+                log.warn("Http call to {} returned a failure response with response {}", httpUrl, serviceNodesResponse);
+                return Optional.empty();
             }
-        } catch (IOException e) {
-            log.error("Error getting service data from the http endPoint: ", e);
-        }
-        return Optional.empty();
-    }
-
-    @Override
-    public boolean isActive() {
-        return true;
+        }, (Function<Exception, Optional<List<ServiceNode<T>>>>) exception -> {
+            log.error("HTTP call to {} returned with an exception. Executing the error handler", httpUrl, exception);
+            return Optional.empty();
+        });
     }
 }