+++ ranger-http/src/main/java/io/appform/ranger/http/servicefinder/HttpShardedServiceFinderBuilder.java
@@ -22,6 +22,7 @@ import io.appform.ranger.core.model.NodeDataSource;
 import io.appform.ranger.core.model.Service;
 import io.appform.ranger.http.config.HttpClientConfig;
 import io.appform.ranger.http.serde.HTTPResponseDataDeserializer;
+import org.apache.hc.client5.http.fluent.Executor;
 
 /**
  *
@@ -30,6 +31,7 @@ public class HttpShardedServiceFinderBuilder<T> extends SimpleShardedServiceFind
 
     private HttpClientConfig clientConfig;
     private ObjectMapper mapper;
+    private Executor httpExecutor;
 
     public HttpShardedServiceFinderBuilder<T> withClientConfig(final HttpClientConfig clientConfig) {
         this.clientConfig = clientConfig;
@@ -41,6 +43,11 @@ public class HttpShardedServiceFinderBuilder<T> extends SimpleShardedServiceFind
         return this;
     }
 
+    public HttpShardedServiceFinderBuilder<T> withHttpExecutor(final Executor httpExecutor){
+        this.httpExecutor = httpExecutor;
+        return this;
+    }
+
     @Override
     public SimpleShardedServiceFinder<T> build() {
         return buildFinder();
@@ -48,7 +55,7 @@ public class HttpShardedServiceFinderBuilder<T> extends SimpleShardedServiceFind
 
     @Override
     protected NodeDataSource<T, HTTPResponseDataDeserializer<T>> dataSource(Service service) {
-        return new HttpNodeDataSource<>(service, clientConfig, mapper);
+        return new HttpNodeDataSource<>(service, clientConfig, mapper, httpExecutor);
     }
 
 }