+++ ranger-http/src/main/java/io/appform/ranger/http/servicefinder/HttpUnshardedServiceFinderBuilider.java
@@ -22,12 +22,14 @@ import io.appform.ranger.core.model.NodeDataSource;
 import io.appform.ranger.core.model.Service;
 import io.appform.ranger.http.config.HttpClientConfig;
 import io.appform.ranger.http.serde.HTTPResponseDataDeserializer;
+import org.apache.hc.client5.http.fluent.Executor;
 
 public class HttpUnshardedServiceFinderBuilider<T>
         extends SimpleUnshardedServiceFinderBuilder<T, HttpUnshardedServiceFinderBuilider<T>, HTTPResponseDataDeserializer<T>> {
 
     private HttpClientConfig clientConfig;
     private ObjectMapper mapper;
+    private Executor httpExecutor;
 
     public HttpUnshardedServiceFinderBuilider<T> withClientConfig(final HttpClientConfig clientConfig) {
         this.clientConfig = clientConfig;
@@ -39,6 +41,11 @@ public class HttpUnshardedServiceFinderBuilider<T>
         return this;
     }
 
+    public HttpUnshardedServiceFinderBuilider<T> withHttpExecutor(final Executor executor){
+        this.httpExecutor = executor;
+        return this;
+    }
+
     @Override
     public SimpleUnshardedServiceFinder<T> build() {
         return buildFinder();
@@ -46,7 +53,7 @@ public class HttpUnshardedServiceFinderBuilider<T>
 
     @Override
     protected NodeDataSource<T, HTTPResponseDataDeserializer<T>> dataSource(Service service) {
-        return new HttpNodeDataSource<>(service, clientConfig, mapper);
+        return new HttpNodeDataSource<>(service, clientConfig, mapper, httpExecutor);
     }
 
 }